Puzzles - Initial System Layout and Barebones Puzzle 1
Adds the general 'puzzles' system with one puzzle so there are versions of everything being used. Includes:

- Puzzle Asset where you provide parameters to the puzzle that are specific to that only.
- Puzzle Generator, the main logic that, given the current state of the game or race, generates the puzzle in a purely data format on a regular grid. This includes the prefabs that should spawn in each cell, and which cells these should 'link' to if necessary.

- IActivator and IActivatee.
IActivator is implemented by scripts on objects that get interacted with by the player in some way (driven over, shot, destroyed) that then need to cause a different object to change.
IActivatee is implemented by scripts that need to recieve this message to do something.
Currently the 'Switch' is an Activator (presumably via being driven over), and that forwards a call to the 'Door' which is an Activatee. Hitting the switch opens the door. It could open more or cause any manner of things, this just keeps it super generic.

- DebugPuzzleSpawner
Once the PuzzleGenerator creates the data form of the puzzle it then needs spawning into the world. This debug spawner does that into world space, eventually there will be some other spawner that does it in Mobius Strip Space.

TODO:
Some sort of manager that selects which puzzle to spawn based on the state of the game. Need the game to be in place first...